+++ sdk/go/aws/provider.go
@@ -19,6 +19,8 @@ type Provider struct {
 	Profile pulumi.StringPtrOutput `pulumi:"profile"`
 	// The region where AWS operations will take place. Examples are `us-east-1`, `us-west-2`, etc.
 	Region pulumi.StringPtrOutput `pulumi:"region"`
+	// The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role for Cloud Control API to use when performing this resource operation. Note, this is a unique feature for server side security enforcement, not to be confused with assumeRole, which is used to obtain temporary client credentials. If you do not specify a role, Cloud Control API uses a temporary session created using your AWS user credentials instead.
+	RoleArn pulumi.StringPtrOutput `pulumi:"roleArn"`
 	// The path to the shared credentials file. If not set this defaults to `~/.aws/credentials`.
 	SharedCredentialsFile pulumi.StringPtrOutput `pulumi:"sharedCredentialsFile"`
 }
@@ -91,6 +93,8 @@ type providerArgs struct {
 	Profile *string `pulumi:"profile"`
 	// The region where AWS operations will take place. Examples are `us-east-1`, `us-west-2`, etc.
 	Region string `pulumi:"region"`
+	// The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role for Cloud Control API to use when performing this resource operation. Note, this is a unique feature for server side security enforcement, not to be confused with assumeRole, which is used to obtain temporary client credentials. If you do not specify a role, Cloud Control API uses a temporary session created using your AWS user credentials instead.
+	RoleArn *string `pulumi:"roleArn"`
 	// Set this to true to force the request to use path-style addressing, i.e., `http://s3.amazonaws.com/BUCKET/KEY`. By default, the S3 client will use virtual hosted bucket addressing when possible (`http://BUCKET.s3.amazonaws.com/KEY`). Specific to the Amazon S3 service.
 	S3ForcePathStyle *bool `pulumi:"s3ForcePathStyle"`
 	// The secret key for API operations. You can retrieve this from the 'Security & Credentials' section of the AWS console.
@@ -135,6 +139,8 @@ type ProviderArgs struct {
 	Profile pulumi.StringPtrInput
 	// The region where AWS operations will take place. Examples are `us-east-1`, `us-west-2`, etc.
 	Region pulumi.StringInput
+	// The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role for Cloud Control API to use when performing this resource operation. Note, this is a unique feature for server side security enforcement, not to be confused with assumeRole, which is used to obtain temporary client credentials. If you do not specify a role, Cloud Control API uses a temporary session created using your AWS user credentials instead.
+	RoleArn pulumi.StringPtrInput
 	// Set this to true to force the request to use path-style addressing, i.e., `http://s3.amazonaws.com/BUCKET/KEY`. By default, the S3 client will use virtual hosted bucket addressing when possible (`http://BUCKET.s3.amazonaws.com/KEY`). Specific to the Amazon S3 service.
 	S3ForcePathStyle pulumi.BoolPtrInput
 	// The secret key for API operations. You can retrieve this from the 'Security & Credentials' section of the AWS console.
@@ -200,6 +206,11 @@ func (o ProviderOutput) Profile() pulumi.StringPtrOutput {
 // The region where AWS operations will take place. Examples are `us-east-1`, `us-west-2`, etc.
 func (o ProviderOutput) Region() pulumi.StringPtrOutput {
 	return o.ApplyT(func(v *Provider) pulumi.StringPtrOutput { return v.Region }).(pulumi.StringPtrOutput)
+}
+
+// The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role for Cloud Control API to use when performing this resource operation. Note, this is a unique feature for server side security enforcement, not to be confused with assumeRole, which is used to obtain temporary client credentials. If you do not specify a role, Cloud Control API uses a temporary session created using your AWS user credentials instead.
+func (o ProviderOutput) RoleArn() pulumi.StringPtrOutput {
+	return o.ApplyT(func(v *Provider) pulumi.StringPtrOutput { return v.RoleArn }).(pulumi.StringPtrOutput)
 }
 
 // The path to the shared credentials file. If not set this defaults to `~/.aws/credentials`.