+++ sdk/nodejs/config/vars.ts
@@ -132,6 +132,17 @@ Object.defineProperty(exports, "region", {
 });
 
 /**
+ * The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role for Cloud Control API to use when performing this resource operation. Note, this is a unique feature for server side security enforcement, not to be confused with assumeRole, which is used to obtain temporary client credentials. If you do not specify a role, Cloud Control API uses a temporary session created using your AWS user credentials instead.
+ */
+export declare const roleArn: string | undefined;
+Object.defineProperty(exports, "roleArn", {
+    get() {
+        return __config.get("roleArn");
+    },
+    enumerable: true,
+});
+
+/**
  * Set this to true to force the request to use path-style addressing, i.e., `http://s3.amazonaws.com/BUCKET/KEY`. By default, the S3 client will use virtual hosted bucket addressing when possible (`http://BUCKET.s3.amazonaws.com/KEY`). Specific to the Amazon S3 service.
  */
 export declare const s3ForcePathStyle: boolean | undefined;