+++ sdk/nodejs/provider.ts
@@ -34,6 +34,10 @@ export class Provider extends pulumi.ProviderResource {
      */
     public readonly region!: pulumi.Output<Region | undefined>;
     /**
+     * The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role for Cloud Control API to use when performing this resource operation. Note, this is a unique feature for server side security enforcement, not to be confused with assumeRole, which is used to obtain temporary client credentials. If you do not specify a role, Cloud Control API uses a temporary session created using your AWS user credentials instead.
+     */
+    public readonly roleArn!: pulumi.Output<string | undefined>;
+    /**
      * The path to the shared credentials file. If not set this defaults to `~/.aws/credentials`.
      */
     public readonly sharedCredentialsFile!: pulumi.Output<string | undefined>;
@@ -63,6 +67,7 @@ export class Provider extends pulumi.ProviderResource {
             resourceInputs["maxRetries"] = pulumi.output(args ? args.maxRetries : undefined).apply(JSON.stringify);
             resourceInputs["profile"] = (args ? args.profile : undefined) ?? utilities.getEnv("AWS_PROFILE");
             resourceInputs["region"] = (args ? args.region : undefined) ?? <any>utilities.getEnv("AWS_REGION", "AWS_DEFAULT_REGION");
+            resourceInputs["roleArn"] = args ? args.roleArn : undefined;
             resourceInputs["s3ForcePathStyle"] = pulumi.output(args ? args.s3ForcePathStyle : undefined).apply(JSON.stringify);
             resourceInputs["secretKey"] = args?.secretKey ? pulumi.secret(args.secretKey) : undefined;
             resourceInputs["sharedCredentialsFile"] = (args ? args.sharedCredentialsFile : undefined) ?? utilities.getEnv("AWS_SHARED_CREDENTIALS_FILE");
@@ -126,6 +131,10 @@ export interface ProviderArgs {
      * The region where AWS operations will take place. Examples are `us-east-1`, `us-west-2`, etc.
      */
     region: pulumi.Input<Region>;
+    /**
+     * The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role for Cloud Control API to use when performing this resource operation. Note, this is a unique feature for server side security enforcement, not to be confused with assumeRole, which is used to obtain temporary client credentials. If you do not specify a role, Cloud Control API uses a temporary session created using your AWS user credentials instead.
+     */
+    roleArn?: pulumi.Input<string>;
     /**
      * Set this to true to force the request to use path-style addressing, i.e., `http://s3.amazonaws.com/BUCKET/KEY`. By default, the S3 client will use virtual hosted bucket addressing when possible (`http://BUCKET.s3.amazonaws.com/KEY`). Specific to the Amazon S3 service.
      */